Renamed files and cleaned up create_taxonomy
