+++ src/graph_extraction/graphautoencoder/.gitignore
@@ -1,0 +1,1 @@
+*.pt
